Update README.md
The system predicts the next word in a sentence based on the input text, aiming to enhance typing efficiency and provide intelligent text suggestions, similar to predictive text used in search engines and messaging apps.  

The dataset used comprises custom sentences, which were preprocessed using tokenization and padding techniques to prepare the data for training. An LSTM model was trained on this dataset for 30 epochs, with continuous improvements in accuracy observed during the training process. The model achieved a training accuracy of 30%, demonstrating its ability to understand and generate text sequences effectively.  

The project was deployed using Streamlit, enabling a user-friendly web interface that provides real-time word predictions as the user types. Recommendations are dynamically updated after each word or space, enhancing the interactive experience.